requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.2
pillow==10.3.0
pytz==2024.1
pywin32==305.1
setuptools==68.2.2
tzdata==2024.1
wheel==0.41.2